requirements.txt
langchain
langgraph
langchain-community
langchain-core
langchain-openai
agno
openai
yfinance
duckduckgo-search
faiss-cpu
sentence-transformers
arxiv
pypdf
python-dotenv
black
langgraph-cli[inmem]
pyppeteer
graphviz
# pygraphviz
mermaid-python